Fix PermissionsEX Group compatibility, possible fix client crashes
Ive added checks to see if the permissions entity being changed is a
group, and to get its players instead of just treating the group as a
player.
I tried to do this to GroupManager, but you can't get a list of group
manager players from a group manager group.
Ive removed removing all players from a team, and just deleting the
team. The server was still stable over many reloads after i removed this
code, and my clients no longer crashed.